Moved CommandLoggers to their own folder
