--- robotbase/src/main/java/com/systemmeltdown/robot/commands/loggers/ClimbCommandsLogger.java
@@ -1,14 +1,0 @@
-package com.systemmeltdown.robot.commands.loggers;
-
-import com.systemmeltdown.robot.subsystems.ClimbSubsystem; // Imported for javadoc (for now)
-
-/**
- * A logging class to log the commands used by the {@link ClimbSubsystem}.
- * 
- * @category Climbing
- * 
- * @category Logging
- */
-class ClimbCommandsLogger {
-    
-}
