+++ robotbase/src/main/java/com/systemmeltdown/robotlog/commands/FeederCommandsLogger.java
@@ -1,4 +1,4 @@
-package com.systemmeltdown.robot.commands.loggers;
+package com.systemmeltdown.robotlog.commands;
 
 import com.systemmeltdown.robot.subsystems.FeederSubsystem; // Imported for javadoc (for now)
