+++ robotbase/src/main/java/com/systemmeltdown/robotlog/commands/ShooterCommandsLogger.java
@@ -1,4 +1,4 @@
-package com.systemmeltdown.robot.commands.loggers;
+package com.systemmeltdown.robotlog.commands;
 
 import com.systemmeltdown.robot.subsystems.ShooterSubsystem; // Imported for javadoc (for now)
